### Add Medusa Server as environment variable
Create a `.env` file in the root of your `storefront` directory. This will store the URL to your Medusa backend server as an environment variable.

```
MEDUSA_URL=http://localhost:9000
```

Update the `nuxt.config.ts` file in the root of your `storefront` directory as follows:

Lastly, update the `pages/index.vue` file so that it uses an environment variable for the URL instead of a hardcoded URL: