### Add "Add to Cart" Button
Update each product in `pages/index.vue` with a button to add a product to the cart.

Confirm whether the products are being added to the cart by performing a curl request as follows:

```bash
curl localhost:9000/store/carts/cart_01HRBTB0X79NAGJYY8T6D5BGK6
```

Replace with the specific `cart id` as listed in your browser console. If all is working your cart should be populated with some products.